se preparan ejemplos para funcionamiento
